refactor: Extract useTransactionExtent
