+++ src/ducks/transactions/TransactionSelectDates.jsx
@@ -1,7 +1,6 @@
-import React, { useState, useEffect, useMemo } from 'react'
+import React, { useEffect, useMemo } from 'react'
 import PropTypes from 'prop-types'
-import { useSelector } from 'react-redux'
-import CozyClient from 'cozy-client'
+
 import SelectDates, { monthRange } from 'components/SelectDates'
 import last from 'lodash/last'
 import uniq from 'lodash/uniq'
@@ -12,15 +11,9 @@ import {
   differenceInCalendarMonths,
   isAfter
 } from 'date-fns'
-import { useQuery, useClient } from 'cozy-client'
 import { getDate } from 'ducks/transactions/helpers'
-import { getFilteringDoc } from 'ducks/filters'
-import { groupsConn, accountsConn } from 'doctypes'
-import {
-  makeFilteredTransactionsConn,
-  makeEarliestLatestQueries
-} from './queries'
-import useSafeState from 'hooks/useSafeState'
+
+import useTransactionExtent from 'hooks/useTransactionExtent'
 
 const rangeMonth = (startDate, endDate) => {
   const options = []
@@ -51,60 +44,6 @@ export const getOptions = transactions => {
       disabled: !mAvailableMonths.has(fmted)
     }
   })
-}
-
-const useConn = conn => {
-  return useQuery(conn.query, conn)
-}
-
-// We do not want queries with the minimal and maximal transaction to receive
-// transactions from other queries
-const extentAutoUpdateOptions = { update: true, add: false, remove: false }
-
-const useTransactionExtent = () => {
-  const client = useClient()
-  const accounts = useConn(accountsConn)
-  const groups = useConn(groupsConn)
-  const [loading, setLoading] = useSafeState(true)
-  const filteringDoc = useSelector(getFilteringDoc)
-  const transactionsConn = makeFilteredTransactionsConn({
-    filteringDoc,
-    accounts,
-    groups
-  })
-  const [data, setData] = useState([])
-
-  useEffect(() => {
-    const fetch = async () => {
-      const baseQuery = transactionsConn.query()
-      const [earliestQuery, latestQuery] = makeEarliestLatestQueries(baseQuery)
-      setLoading(true)
-      try {
-        const [earliest, latest] = await Promise.all(
-          [earliestQuery, latestQuery].map(async (q, i) => {
-            const queryName = `${transactionsConn.as}-${
-              i === 0 ? 'earliest' : 'latest'
-            }`
-            await client.query(q, {
-              fetchPolicy: CozyClient.fetchPolicies.olderThan(30 * 1000),
-              as: queryName,
-              autoUpdate: extentAutoUpdateOptions
-            })
-            return client.getQueryFromState(queryName)
-          })
-        )
-        setData([earliest.data[0], latest.data[0]])
-      } finally {
-        setLoading(false)
-      }
-    }
-
-    if (transactionsConn.enabled) {
-      fetch()
-    }
-  }, [transactionsConn.enabled, transactionsConn.as]) // eslint-disable-line
-
-  return [data[0], data[1], loading]
 }
 
 const getMonthFromTransaction = transaction => {
